Contains the new machinery for LHAPDF5 or LHAPDF6 plugins. You will find Phil's description below. I have checked that there should be no clashes with other changes since 308. The Update History has not yet been updated.
Since I have been gone over the weekend, and start teaching tomorrow,
and have to move office this week, and anyway have problems with 
installing LHAPDF on my Mac, it is not clear that I will have time 
to run any checks until Wednesday. The rest of you, in particular 
Steve and Stefan, please go ahead, so we can discuss it then.

Best Torbjörn

--------------------------------

From Phil:

Dear Torbjörn and Steve,

I now have version of Pythia 8.2 working with LHAPDF5 and LHAPDF6 in a
fashion that I am happy with, and that has been tested on both Linux
and OS X. Here are the salient points:

* The LHAPDF class interface is identical to the old interface.
However it needs an Info pointer to load a PDF, since this requires
global handling of what plugin libraries have been loaded (this is all
handled relatively behind the scene and is not actually very
complicated).
* The PDFs for beam A and beam B can be set independently (main, hard,
and pion).
* I have updated all relevant documentation (xml and README's).
* I have updated all relevant examples.
* I have updated all relevant configuration and Make files.

Ok, so going into a bit more detail on the implementation.

* The major work has been done in PartonDistributions.cc, LHAPDF5.h,
and LHAPDF6.h.
* The LHAPDF5.h and LHAPDF6.h source allows for the creation of a
runtime plugin class that is loaded by the LHAPDF class, defined in
PartonDistributions.cc.
* Pythia no longer needs to link against any type of LHAPDF library,
since this is now loaded at runtime. This is nice, because now Pythia
is truly external package dependency free.

For the examples I have the following comments.

* main54: Does not work anymore because these sets do not exist any
longer. Do we want to remove the example? We can use similar sets, but
they don't reach the required precision of the example.
* main61: Identical to main42 except it uses LHAPDF. Since LHAPDF no
longer needs to be linked, do we want to remove this example? Migrated
the additional options to main42.cmnd and removed this example for
now.
* main62: Maybe this should be moved to main43 for consistency? Moved for now.
* main8[1-8]: These did not compile (but not because of LHAPDF
issues). I've now modified them to compile, added example usage, and
changed the necessary LHAPDF commands in main8[1-8].cmnd. I've also
added main83.cmnd (since it seems like it should really be there).
These should probably all be checked by Stefan.

I've attached "lhapdf.diff" which gives the svn differences between
what I have, and revision 308 of Pythia. I see that now revision 314
is available, so hopefully not too many changes have gone in that
might conflict with these changes.

To apply the patch, just navigate to your SVN copy of pythia82/trunk
and run "patch -p0 -i lhapdf.diff". After patching, configure with
something like "./configure --with-lhapdf5=/path/to/lhapdf5" and then
"make". Running "main51" should then hopefully work without issue.

I realize this is a large patch, with a fair number of changes, but I
think that the ones that don't concern documentation are actually
minimal. Please let me know what you think and have a great weekend!

Best,

- Phil



git-svn-id: svn+ssh://svn.hepforge.org/hepforge/svn/pythia8/pythia82/trunk@315 ffdf98f5-c1f1-482d-8be8-39a3f79b3492