From Phil: (1) I've changed all the compile statements to handle the issues that Steve and Jesper were having. Specifically everything now goes, <compiler> <in> <out> <include> <flags> <libraries>, which I think should take care of theses issues. (2) I've inlcuded the -lgfortran for the LHAPDF5 libraries. This is only needed when the static version of the LHAPDF5 library is being used, but it is fine to include it either way. (3) I've updated the configure script to handle the issue that Peter was seeing with Clang. (4) I've updated the pythia8-config script, which should be placed in bin. The file pythia8-config.in is no longer needed. This should handle all old behaviour, but it also handles any of the standard configuration arguments as well. I've updated the configure and pythia8-config scripts to now handle a -h, -help, --h, or --help anywhere in the arguments list. It now does a first pass looking for these, and if it does exits after printing the help message. Tiny bug fix in tau decays.
From Steve:
The first request was to use  "namespace Pythia8 { ... }" instead of
"using namespace Pythia8;"
The next was to "inline" all of the extensive implementations in the
the Pythia8Plugins header files.
The third was to reallow access in Event to the list of daughters
of a particle, since this is currently how we use Pythia8 with Evtgen.
This is just a stop-gap request until Evtgen can be updated.  I'm not
sure I did this correctly, so please check.
The fourth was to make the access to the SLHAheader public instead
of private (since we read and connect our LHE files and slha info to
Pythia8 in our own way).

From Stefan: correct typo in ProcessContainer.cc.

Some documentation updates.

One tiny bug fix from Stefan still to come, for the rest no
known problems to fix. We are getting there...



git-svn-id: svn+ssh://svn.hepforge.org/hepforge/svn/pythia8/pythia82/trunk@325 ffdf98f5-c1f1-482d-8be8-39a3f79b3492