* CJKL photon PDFs, first part of gamma-gamma machinery.   Ilkka listed as co-author. * Further checks junction topologies, mainly for collinear partons. * MBR diffraction renormalization option. * Move up random-number initialization. * Minor update beam-remnant handling for DIS.
A reminder: when you send code updates, do it based on the most recent
version in svn. Many of you have your own private working versions
for development, which is as it should be. But if then you never sync 
them with what is changed in svn, least of all in the methods you work
most on, you never notice that other changes may have occured in the
meantime. So I have to reintroduce changed code over and over again. 



git-svn-id: svn+ssh://svn.hepforge.org/hepforge/svn/pythia8/pythia82/trunk@427 ffdf98f5-c1f1-482d-8be8-39a3f79b3492