A new switch, TimeShower:MEextended, on by defaults, attempts to guess the most relevant ME correction when the correct choice is not known or implemented.
Increased possibilities to set the epsilon and alpha' parameters 
of the Pomeron trajectory for hard-diffraction Pomeron fluxes. 
 
Extrapolation of PDFs to small x values when PDF:extrapolate = on
now extended to more cases. 
 
New flag TimeShower:QEDshowerByOther allows charged resonances, 
like the W^+-, to radiate photons.

The description of POWHEG:vetoCount = 0 has been corrected. 

Allow for lightest neutralino not to decay as a resonance.
 
Minor further updates in LesHouches.h and in manual.



git-svn-id: svn+ssh://svn.hepforge.org/hepforge/svn/pythia8/pythia82/trunk@441 ffdf98f5-c1f1-482d-8be8-39a3f79b3492