* Remove many rarely (if ever) used ostream& os = cout optional arguments in favour of hardcoded cout in the code. Eliminates some redundancy of methods.
* Rename ...print(... methods to ...list(... to favour a more regular 
naming pattern. 

* Make some SlowJet methods virtual to allow derived classes with 
modified properties. 
 
* Moved some misplaced info on parton-level choice of MPIs. 
 
* Corrected typos where some bottomonium long-distance matrix element 
had been set larger than normally assumed.

* Corrected default (unused) ../xmldoc/Index.xml to 
../share/Pythia8/xmldoc/Index.xml.

---------------------------------------------------------------------

The current batch of changes spans many files, but is mainly
straightforward in nature. The elimination of many unused 
ostream& os = cout was overdue. At the same time Phil's complaint
about irregular use of list() vs print() is now resolved in favour of 
list(), which is what is used for the key methods like event.list().
Some irregularities remain, e.g. when cout would no be the natural
destination. 

Phil (and everybody else, of course): do have a look if there are 
further irregularities, to begin with on the printing front, and also
if I have gone too far somewhere.

Steve: your changes should be in now. For some reason patch -i did 
not work although it has worked earlier; it claimed it could not find 
the files although the path it showed seemed correct. Anyway, not so 
many changes so easy to do manually.



git-svn-id: svn+ssh://svn.hepforge.org/hepforge/svn/pythia8/pythia82/trunk@447 ffdf98f5-c1f1-482d-8be8-39a3f79b3492