Dear All,
We now at long last are where we were supposed to have been two weeks 
ago, with respect to the next release. It has been a very long and
tiring chase to make things work. In view of this we may also need
an extra long period to do checks that all now is well, or at least
no worse than it used to be, say a month ago. In your continued tests
and corrections suggestions, do use this new svn version as baseline.

Best, Torbjörn

Changes in this version: 
- Uncertainty band variations code and documentation now in place 
  and working (hopefully).
- Corrected cross-section bookkeeping/calculation.
- A few stability updates for the e+e- -> gamma-gamma machinery.
- Fixed harmless out-of-bounds in HelicityMatrixElements.



git-svn-id: svn+ssh://svn.hepforge.org/hepforge/svn/pythia8/pythia82/trunk@484 ffdf98f5-c1f1-482d-8be8-39a3f79b3492