Bug fix the calculation of enhancement factor for the machinery with two hard processes.
New method Settings::getReadHistory returns a vector with all strings 
that have been read in by the Settings::readString()</code> method. 
(Still to be expanded?) 
 
When using the PhaseSpace:bias2Selection to reweight high-pT events 
the PhaseSpace:pTHatMinDiverge value is now used as ultimate fail-safe 
to avoid pT = 0. 
 
A few minor fixes when Dark Matter is used as incoming beams in 
Les Houches input. 
 
Fixed memory leak in TimeShower, for the case when ProcessLevel is off. 
 
Minor numerical precision improvements in gamma-gamma kinematics. 



git-svn-id: svn+ssh://svn.hepforge.org/hepforge/svn/pythia8/pythia82/trunk@550 ffdf98f5-c1f1-482d-8be8-39a3f79b3492